backend/requirements.txt
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
redis==7.0.0
requests==2.32.5
soupsieve==2.8
SQLAlchemy==2.0.44
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
pandas
numpy==1.26.4
scikit-surprise==1.1.3
nltk
scikit-learn